requirements.txt
numpy
pandas
seaborn
langchain_community
jupyter
langchain-text-splitters
python-dotenv
langchain-openai
langchain-huggingface
tqdm
langchain_ollama
faiss-cpu
langchain_chroma
langchain_openai
langchain_core
streamlit
groq
langchain_groq

fastapi
uvicorn
langserve
sse_starlette
pydantic==1.10.13